Test suite can access private parameter methods
Most of the actual parameter tests don't even involve a parameter set, so
calling setters for parameters is of no harm in these cases. However, rather
than exposing some variant of the set function, the test runner is now a
friend class of the parameter (ugh @ friend classes, but it works).