Add method to give thread low priority
Sometimes when creating a background thread it's useful to tell the
system that it should run with a lower scheduling priority. However,
creating a more flexible method which would take in a priority level
is a bad idea, because there's no use in setting a thread to the
default priority and scheduling to increased priority is usually a bad
idea.

Therefore this commit takes the simple approach and simply offers to
decrease the priority, which should usually work.