Rename PluginParametersTests to _Tests
Having this friend class is needed because when ENABLE_MULTITHREADED
is true, then setters of PluginParameter methods are no longer public.
So to keep these tests mostly the same for both versions of
PluginParameters, we allow access to protected methods via a friend
class, which is enabled by the preprocessor when building the test
suite.