Discourage scheduling events directly after construction
It is extremely difficult to ensure that the async thread is actually
waiting on the condition variable when notify() is sent. Actually it's
rather easy to ensure this, however that solution would involve a
mutex around scheduleEvent(), which could result in priority
inversion.

For this reason, the complex (and not-completely working) guarantee of
synchronous construction is removed, as it cannot be delivered.