Events always dispatched to realtime thread
Although it would be nice to immediately alert observers on the same
thread before re-dispatching, this unfortunately will not work in the
case of events sent by the async thread. This is because the event's
apply() method is *only* called by the main thread, which is a hard
requirement or else multiple events for the same parameter could cause
the final value to be incorrect.

In this change, async events now dispatch to the realtime thread right
away. In the realtime thread, the priority is always inverted and the
event is re-dispatched right away.