Add status bar component
By default, listens to all parameters which have at least one
PluginParameterComponent observing them and are not booleans. Also added hooks
to display the current value of a parameter on mouseover (only applied to
knobs and faders).

Also added a nice little fade-out effect when the text is cleared from the
screen. :)