StatusBar can subscribe to relevant parameters
Rather than subscribing to all parameters in the set, a method is now
added to allow the status bar to find parameters and subscribe them
selectively. This improves the performance of onParameterUpdated(), and
also requires only one method to be called in the GUI constructor.

Additionally, this change fixes an issue where the order of
initialization would cause bugs if the StatusBar were not the last
component added.